addressed one set of comments
