+++ core/domain/exp_services.py
@@ -824,9 +824,7 @@ def _save_exploration(committer_id, exploration, commit_message, change_list):
     exploration.version += 1
 
     # Trigger statistics model update.
-    taskqueue_services.defer(
-        stats_services.handle_stats_creation,
-        taskqueue_services.QUEUE_NAME_DEFAULT, exploration, change_list)
+    stats_services.handle_stats_creation(exploration, change_list)
 
     if feconf.ENABLE_ML_CLASSIFIERS:
         new_to_old_state_names = exploration.get_state_names_mapping(
@@ -889,10 +887,7 @@ def _create_exploration(
     exploration.version += 1
 
     # Trigger statistics model creation.
-    taskqueue_services.defer(
-        stats_services.handle_new_exp_stats_creation,
-        taskqueue_services.QUEUE_NAME_DEFAULT,
-        exploration)
+    stats_services.handle_new_exp_stats_creation(exploration)
 
     if feconf.ENABLE_ML_CLASSIFIERS:
         # Find out all states that need a classifier to be trained.
