+++ core/domain/stats_services_test.py
@@ -22,7 +22,6 @@ from core.domain import exp_services
 from core.domain import stats_jobs_continuous
 from core.domain import stats_domain
 from core.domain import stats_services
-from core.domain import stats_services_old
 from core.domain import user_services
 from core.platform import models
 from core.tests import test_utils
@@ -44,9 +43,10 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
         self.stats_model_id = (
             stats_models.ExplorationStatsModel.create('exp_id1', 1, 0, 0, {}))
 
-    def test_async_call_to_stats_methods(self):
-        """Test that async calls are being made to the handle_stats_creation and
-        handle_new_exp_stats_creation methods.
+    def test_calls_to_stats_methods(self):
+        """Test that calls are being made to the handle_stats_creation and
+        handle_new_exp_stats_creation methods when an exploration is created or
+        updated.
         """
         # Create exploration object in datastore.
         exp_id = 'exp_id'
@@ -60,23 +60,10 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
                 assets_list)
         exploration = exp_services.get_exploration_by_id(exp_id)
 
-        # Now, there is one async call to handle_new_exp_stats_creation in the
-        # queue.
-        self.assertEqual(
-            self.count_jobs_in_taskqueue(taskqueue_services.QUEUE_NAME_DEFAULT),
-            1)
-        with self.assertRaisesRegexp(
-            Exception,
-            'Entity for class ExplorationStatsModel with id %s not found' % (
-                exploration.id + '.' + str(exploration.version))):
-            stats_services.get_exploration_stats_by_id(
-                exploration.id, exploration.version)
-
-        # Process the task to create stats model.
-        self.process_and_flush_pending_tasks()
-        self.assertEqual(
-            self.count_jobs_in_taskqueue(taskqueue_services.QUEUE_NAME_DEFAULT),
-            0)
+        # Now, there are two analytics model in datastore (one from setup
+        # function).
+        all_models = stats_models.ExplorationStatsModel.get_all()
+        self.assertEqual(all_models.count(), 2)
         exploration_stats = stats_services.get_exploration_stats_by_id(
             exploration.id, exploration.version)
         self.assertEqual(
@@ -92,23 +79,10 @@ class StatisticsServicesTest(test_utils.GenericTestBase):
         exp_services.update_exploration(
             feconf.SYSTEM_COMMITTER_ID, exp_id, change_list, '')
 
-        # Now, there should be one async call to handle_stats_creation in the
-        # queue.
-        self.assertEqual(
-            self.count_jobs_in_taskqueue(taskqueue_services.QUEUE_NAME_DEFAULT),
-            1)
-        with self.assertRaisesRegexp(
-            Exception,
-            'Entity for class ExplorationStatsModel with id %s not found' % (
-                exploration.id + '.' + str(exploration.version))):
-            stats_services.get_exploration_stats_by_id(
-                exploration.id, exploration.version)
-
-        # Process the task to create stats model.
-        self.process_and_flush_pending_tasks()
-        self.assertEqual(
-            self.count_jobs_in_taskqueue(taskqueue_services.QUEUE_NAME_DEFAULT),
-            0)
+        # Now, there should be three analytics models in datastore (one from
+        # setup function).
+        all_models = stats_models.ExplorationStatsModel.get_all()
+        self.assertEqual(all_models.count(), 3)
         exploration_stats = stats_services.get_exploration_stats_by_id(
             exploration.id, exploration.version)
         self.assertEqual(
@@ -363,7 +337,7 @@ class AnswerEventTests(test_utils.GenericTestBase):
         exp_version = exp.version
 
         for state_name in [first_state_name, second_state_name]:
-            state_answers = stats_services_old.get_state_answers(
+            state_answers = stats_services.get_state_answers(
                 'eid', exp_version, state_name)
             self.assertEqual(state_answers, None)
 
@@ -449,19 +423,19 @@ class AnswerEventTests(test_utils.GenericTestBase):
             'session_id': 'sid5', 'interaction_id': 'Continue', 'params': {}
         }]
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             'eid', exp_version, first_state_name)
         self.assertEqual(
             state_answers.get_submitted_answer_dict_list(),
             expected_submitted_answer_list1)
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             'eid', exp_version, second_state_name)
         self.assertEqual(
             state_answers.get_submitted_answer_dict_list(),
             expected_submitted_answer_list2)
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             'eid', exp_version, third_state_name)
         self.assertEqual(
             state_answers.get_submitted_answer_dict_list(),
@@ -482,7 +456,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
             self.EXP_ID, self.owner_id, end_state_name='End')
 
     def test_record_answer_without_retrieving_it_first(self):
-        stats_services_old.record_answer(
+        stats_services.record_answer(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             stats_domain.SubmittedAnswer(
@@ -490,7 +464,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 0, exp_domain.EXPLICIT_CLASSIFICATION, {},
                 'a_session_id_val', 1.0))
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
@@ -505,12 +479,12 @@ class RecordAnswerTests(test_utils.GenericTestBase):
         }])
 
     def test_record_and_retrieve_single_answer(self):
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertIsNone(state_answers)
 
-        stats_services_old.record_answer(
+        stats_services.record_answer(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             stats_domain.SubmittedAnswer(
@@ -518,7 +492,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 1, exp_domain.EXPLICIT_CLASSIFICATION, {},
                 'a_session_id_val', 10.0))
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(state_answers.exploration_id, 'exp_id0')
@@ -538,7 +512,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
         }])
 
     def test_record_and_retrieve_single_answer_with_preexisting_entry(self):
-        stats_services_old.record_answer(
+        stats_services.record_answer(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             stats_domain.SubmittedAnswer(
@@ -546,7 +520,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 0, exp_domain.EXPLICIT_CLASSIFICATION, {},
                 'a_session_id_val', 1.0))
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
@@ -560,7 +534,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
             'params': {}
         }])
 
-        stats_services_old.record_answer(
+        stats_services.record_answer(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             stats_domain.SubmittedAnswer(
@@ -568,7 +542,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 1, exp_domain.EXPLICIT_CLASSIFICATION, {},
                 'a_session_id_val', 10.0))
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(state_answers.exploration_id, 'exp_id0')
@@ -597,7 +571,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
         }])
 
     def test_record_many_answers(self):
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertIsNone(state_answers)
@@ -613,13 +587,13 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 'answer bbbbb', 'TextInput', 1, 0,
                 exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
         ]
-        stats_services_old.record_answers(
+        stats_services.record_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             submitted_answer_list)
 
         # The order of the answers returned depends on the size of the answers.
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(state_answers.exploration_id, 'exp_id0')
@@ -662,7 +636,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
         with self.swap(
             stats_models.StateAnswersModel, '_MAX_ANSWER_LIST_BYTE_SIZE',
             100000):
-            state_answers = stats_services_old.get_state_answers(
+            state_answers = stats_services.get_state_answers(
                 self.EXP_ID, self.exploration.version,
                 self.exploration.init_state_name)
             self.assertIsNone(state_answers)
@@ -681,7 +655,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                     exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v',
                     7.5),
             ]
-            stats_services_old.record_answers(
+            stats_services.record_answers(
                 self.EXP_ID, self.exploration.version,
                 self.exploration.init_state_name, 'TextInput',
                 submitted_answer_list * 200)
@@ -695,7 +669,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
 
             # The order of the answers returned depends on the size of the
             # answers.
-            state_answers = stats_services_old.get_state_answers(
+            state_answers = stats_services.get_state_answers(
                 self.EXP_ID, self.exploration.version,
                 self.exploration.init_state_name)
             self.assertEqual(state_answers.exploration_id, 'exp_id0')
@@ -707,7 +681,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 len(state_answers.get_submitted_answer_dict_list()), 600)
 
     def test_record_many_answers_with_preexisting_entry(self):
-        stats_services_old.record_answer(
+        stats_services.record_answer(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             stats_domain.SubmittedAnswer(
@@ -715,7 +689,7 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 0, exp_domain.EXPLICIT_CLASSIFICATION, {},
                 'a_session_id_val', 1.0))
 
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(state_answers.get_submitted_answer_dict_list(), [{
@@ -740,13 +714,13 @@ class RecordAnswerTests(test_utils.GenericTestBase):
                 'answer bbbbbbb', 'TextInput', 1, 0,
                 exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
         ]
-        stats_services_old.record_answers(
+        stats_services.record_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             submitted_answer_list)
 
         # The order of the answers returned depends on the size of the answers.
-        state_answers = stats_services_old.get_state_answers(
+        state_answers = stats_services.get_state_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(state_answers.exploration_id, 'exp_id0')
@@ -819,12 +793,12 @@ class SampleAnswerTests(test_utils.GenericTestBase):
                 exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
         ]
         # Record 600 answers.
-        stats_services_old.record_answers(
+        stats_services.record_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             submitted_answer_list * 200)
 
-        sample_answers = stats_services_old.get_sample_answers(
+        sample_answers = stats_services.get_sample_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(len(sample_answers), 100)
@@ -836,12 +810,12 @@ class SampleAnswerTests(test_utils.GenericTestBase):
                 exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 10.0)
         ]
         # Record 100 answers.
-        stats_services_old.record_answers(
+        stats_services.record_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             submitted_answer_list * 100)
 
-        sample_answers = stats_services_old.get_sample_answers(
+        sample_answers = stats_services.get_sample_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(sample_answers, ['answer a'] * 100)
@@ -856,12 +830,12 @@ class SampleAnswerTests(test_utils.GenericTestBase):
                 exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v', 7.5),
         ]
         # Record 2 answers.
-        stats_services_old.record_answers(
+        stats_services.record_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name, 'TextInput',
             submitted_answer_list)
 
-        sample_answers = stats_services_old.get_sample_answers(
+        sample_answers = stats_services.get_sample_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertEqual(sample_answers, ['answer a', 'answer bbbbb'])
@@ -872,7 +846,7 @@ class SampleAnswerTests(test_utils.GenericTestBase):
         with self.swap(
             stats_models.StateAnswersModel, '_MAX_ANSWER_LIST_BYTE_SIZE',
             15000):
-            state_answers = stats_services_old.get_state_answers(
+            state_answers = stats_services.get_state_answers(
                 self.EXP_ID, self.exploration.version,
                 self.exploration.init_state_name)
             self.assertIsNone(state_answers)
@@ -883,7 +857,7 @@ class SampleAnswerTests(test_utils.GenericTestBase):
                     exp_domain.EXPLICIT_CLASSIFICATION, {}, 'session_id_v',
                     3.0),
             ]
-            stats_services_old.record_answers(
+            stats_services.record_answers(
                 self.EXP_ID, self.exploration.version,
                 self.exploration.init_state_name, 'TextInput',
                 submitted_answer_list * 100)
@@ -899,7 +873,7 @@ class SampleAnswerTests(test_utils.GenericTestBase):
 
         # Verify that the list of sample answers returned contains fewer than
         # 100 answers, although a total of 100 answers were submitted.
-        sample_answers = stats_services_old.get_sample_answers(
+        sample_answers = stats_services.get_sample_answers(
             self.EXP_ID, self.exploration.version,
             self.exploration.init_state_name)
         self.assertLess(len(sample_answers), 100)
